pyproject.toml
[tool.poetry]
name = "vigilant-crypto-snatch"
version = "3.0.3"
description = "Crypto currency buying agent"
authors = ["Martin Ueding <mu@martin-ueding.de>"]
license = "MIT"

[tool.poetry.scripts]
vigilant-crypto-snatch = "vigilant_crypto_snatch.cli:main"

[tool.poetry.dependencies]
BitstampClient = "^2.2.8"
appdirs = "^1.4.4"
clikraken = "^0.8.3"
coloredlogs = "^15.0"
pandas = { version = "^1.2.3", optional = true }
python = "^3.7.1,<3.10"
pyyaml = "^5.4.1"
requests = "^2.25.1"
scipy = {version = "^1.6.2", optional = true}
sqlalchemy = "^1.3.23"
streamlit = {version = "^0.80.0", optional = true}
urllib3 = "^1.26.3"

[tool.poetry.extras]
evaluation = ["pandas", "scipy", "streamlit"]

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mkdocs_material = "^7.0.5"
pytest = "^6.2.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
